require JDK 8; add util class for working with java.util.stream.Stream
