+++ inmemory/src/main/java/org/vertexium/inmemory/InMemoryStreamingPropertyValueRef.java
@@ -3,7 +3,7 @@ package org.vertexium.inmemory;
 import org.vertexium.VertexiumException;
 import org.vertexium.property.StreamingPropertyValue;
 import org.vertexium.property.StreamingPropertyValueRef;
-import org.vertexium.util.StreamUtils;
+import org.vertexium.util.IOUtils;
 
 import java.io.IOException;
 
@@ -13,7 +13,7 @@ class InMemoryStreamingPropertyValueRef extends StreamingPropertyValueRef<InMemo
     InMemoryStreamingPropertyValueRef(StreamingPropertyValue value) {
         super(value);
         try {
-            this.valueData = StreamUtils.toBytes(value.getInputStream());
+            this.valueData = IOUtils.toBytes(value.getInputStream());
         } catch (IOException ex) {
             throw new VertexiumException("Could not read streaming property value", ex);
         }
